APK and relavent source files
The app is very simple. It consists of a button which, when pressed will cause a dice to appear on the screen and be "rolled" to the proper value (no animations, just still images being swapped out).

Source files included are the only files with changes from what Android Studio generates when ran. The changes made are in accordance with the video mentioned in the repository name. I'll experiment more later on, but for now am working alongside the tutorials.